Brings infrastructure fixes and Ladder Script implementation onto main.
Ladder Script C++ will be removed from ghost-core in the next commit
(it belongs in the ghost-labs-ladder-script repo).

Infrastructure fixes included:
- Tree sync vs primary-supersedes-fallback race condition
- Checkpoint pipeline stall fix
- Phase17 config restore
- Qt build fix (int64_t cast)
- Nullifier crash recovery, schema v30
- Ghost-pay restart in chaos tests

Merge conflicts resolved:
- nullifier_route_handler.rs: kept canonical_root fix from main
- ssh.rs: kept ghost-pay restart from feature, drop-in dir from main
- ladder-engine/index.html: took feature version (newer)
- ghost-labs.html: took feature version (newer)

+++ ghost-core/src/qt/createlockwizard.cpp
@@ -419,7 +419,7 @@ bool CreateLockConfirmPage::validatePage()
         request.setHeader(QNetworkRequest::ContentTypeHeader, QStringLiteral("application/json"));
 
         QJsonObject body;
-        body[QStringLiteral("amount_sats")] = GhostPay::denominationSats(wiz->selectedDenomination());
+        body[QStringLiteral("amount_sats")] = static_cast<double>(GhostPay::denominationSats(wiz->selectedDenomination()));
         body[QStringLiteral("timelock_tier")] = static_cast<int>(wiz->selectedTimelockTier());
         if (!wiz->lockLabel().isEmpty()) {
             body[QStringLiteral("label")] = wiz->lockLabel();
